requirements.txt
RPi.GPIO
blynklib
boto3
